Move OpenAI subscription code to separate crate
