+++ crates/openai_subscribed/Cargo.toml
@@ -1,0 +1,36 @@
+[package]
+name = "openai_subscribed"
+version = "0.1.0"
+edition.workspace = true
+publish.workspace = true
+license = "GPL-3.0-or-later"
+
+[lints]
+workspace = true
+
+[lib]
+path = "src/openai_subscribed.rs"
+
+[dependencies]
+anyhow.workspace = true
+base64.workspace = true
+credentials_provider.workspace = true
+futures.workspace = true
+gpui.workspace = true
+http_client.workspace = true
+language_model.workspace = true
+log.workspace = true
+oauth_callback_server.workspace = true
+open_ai.workspace = true
+rand.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+sha2.workspace = true
+smol.workspace = true
+url.workspace = true
+util.workspace = true
+
+[dev-dependencies]
+gpui = { workspace = true, features = ["test-support"] }
+http_client = { workspace = true, features = ["test-support"] }
+parking_lot.workspace = true
